# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Fakebook Social Simulator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  name: string;
  username: string;
  avatarUrl: string;
  coverUrl: string;
  /** @nullable */
  bio: string | null;
  friendCount: number;
  postCount: number;
  isCurrentUser: boolean;
  /** @nullable */
  friendshipStatus?: string | null;
}

export interface Reactions {
  like?: number;
  love?: number;
  haha?: number;
  wow?: number;
  sad?: number;
  angry?: number;
}

export interface Post {
  id: number;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  userId: number;
  userName: string;
  userAvatarUrl: string;
  likeCount: number;
  commentCount: number;
  shareCount: number;
  liked: boolean;
  /** @nullable */
  userReaction?: string | null;
  reactions: Reactions;
  createdAt: string;
}

export interface PostInput {
  /** @minLength 1 */
  content: string;
  imageUrl?: string;
}

export interface Comment {
  id: number;
  postId: number;
  userId: number;
  userName: string;
  userAvatarUrl: string;
  content: string;
  /** @nullable */
  parentId?: number | null;
  createdAt: string;
  replies?: Comment[];
}

export interface CommentInput {
  /** @minLength 1 */
  content: string;
  parentId?: number;
}

export interface LikeInput {
  reactionType?: string;
}

export interface LikeResult {
  liked: boolean;
  likeCount: number;
  /** @nullable */
  userReaction?: string | null;
  reactions: Reactions;
}

export interface FriendRequest {
  id: number;
  fromUserId: number;
  toUserId: number;
  fromUserName: string;
  fromUserAvatarUrl: string;
  status: string;
  createdAt: string;
}

export interface FriendList {
  friends: User[];
  pendingRequests: FriendRequest[];
}

export interface FriendRequestInput {
  toUserId: number;
}

export interface FriendResponseInput {
  action: string;
}

export interface Notification {
  id: number;
  userId: number;
  type: string;
  message: string;
  read: boolean;
  /** @nullable */
  relatedPostId?: number | null;
  /** @nullable */
  relatedUserId?: number | null;
  /** @nullable */
  relatedUserName?: string | null;
  /** @nullable */
  relatedUserAvatarUrl?: string | null;
  createdAt: string;
}

export interface FeedSummary {
  totalPosts: number;
  totalLikes: number;
  totalComments: number;
  unreadNotifications: number;
  unreadMessages: number;
  onlineFriendsCount: number;
  /** @nullable */
  trendingPostId: number | null;
}

export interface Conversation {
  id: number;
  otherUser: User;
  /** @nullable */
  lastMessage?: string | null;
  /** @nullable */
  lastMessageSenderId?: number | null;
  unreadCount: number;
  createdAt: string;
  lastMessageAt: string;
}

export interface ConversationInput {
  toUserId: number;
}

export interface Message {
  id: number;
  conversationId: number;
  senderId: number;
  senderName: string;
  senderAvatarUrl: string;
  content: string;
  createdAt: string;
  read: boolean;
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
}

export type ListPostsParams = {
userId?: number;
};

export type MarkConversationRead200 = {
  ok?: boolean;
};