# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Fakebook Social Simulator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: Get feed posts
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{postId}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a single post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /posts/{postId}/like:
    post:
      operationId: toggleLike
      tags: [posts]
      summary: Toggle reaction on a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LikeInput"
      responses:
        "200":
          description: Updated like status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResult"
  /posts/{postId}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: Get comments for a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Add a comment to a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /posts/{postId}/comments/{commentId}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /friends:
    get:
      operationId: listFriends
      tags: [friends]
      summary: List friends and pending requests for current user
      responses:
        "200":
          description: Friends list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendList"
    post:
      operationId: sendFriendRequest
      tags: [friends]
      summary: Send a friend request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FriendRequestInput"
      responses:
        "201":
          description: Friend request sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendRequest"
  /friends/{requestId}:
    patch:
      operationId: respondFriendRequest
      tags: [friends]
      summary: Accept or reject a friend request
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FriendResponseInput"
      responses:
        "200":
          description: Updated friend request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendRequest"
...
  /notifications/{notificationId}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, username, avatarUrl, coverUrl, bio, friendCount, postCount, isCurrentUser]
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        avatarUrl:
          type: string
        coverUrl:
          type: string
        bio:
          type: ["string", "null"]
        friendCount:
          type: integer
        postCount:
          type: integer
        isCurrentUser:
          type: boolean
        friendshipStatus:
          type: ["string", "null"]
    Reactions:
      type: object
      properties:
        like:
          type: integer
        love:
          type: integer
        haha:
          type: integer
        wow:
          type: integer
        sad:
          type: integer
        angry:
          type: integer
    Post:
      type: object
      required: [id, content, userId, userName, userAvatarUrl, likeCount, commentCount, shareCount, liked, createdAt, reactions]
      properties:
        id:
          type: integer
        content:
          type: string
        imageUrl:
          type: ["string", "null"]
        userId:
          type: integer
        userName:
          type: string
        userAvatarUrl:
          type: string
        likeCount:
          type: integer
        commentCount:
          type: integer
        shareCount:
          type: integer
        liked:
          type: boolean
        userReaction:
          type: ["string", "null"]
        reactions:
          $ref: "#/components/schemas/Reactions"
        createdAt:
          type: string
    PostInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
        imageUrl:
          type: string
    Comment:
      type: object
      required: [id, postId, userId, userName, userAvatarUrl, content, createdAt]
      properties:
        id:
          type: integer
        postId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userAvatarUrl:
          type: string
        content:
          type: string
        parentId:
          type: ["integer", "null"]
        createdAt:
          type: string
        replies:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
    CommentInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
        parentId:
          type: integer
    LikeInput:
      type: object
      properties:
        reactionType:
          type: string
    LikeResult:
      type: object
      required: [liked, likeCount, reactions]
      properties:
        liked:
          type: boolean
        likeCount:
          type: integer
        userReaction:
          type: ["string", "null"]
        reactions:
          $ref: "#/components/schemas/Reactions"
    FriendRequest:
      type: object
      required: [id, fromUserId, toUserId, fromUserName, fromUserAvatarUrl, status, createdAt]
      properties:
        id:
          type: integer
        fromUserId:
          type: integer
        toUserId:
          type: integer
        fromUserName:
          type: string
        fromUserAvatarUrl:
          type: string
        status:
          type: string
        createdAt:
          type: string
    FriendList:
      type: object
      required: [friends, pendingRequests]
      properties:
        friends:
          type: array
          items:
            $ref: "#/components/schemas/User"
        pendingRequests:
          type: array
          items:
            $ref: "#/components/schemas/FriendRequest"
    FriendRequestInput:
      type: object
      required: [toUserId]
      properties:
        toUserId:
          type: integer
    FriendResponseInput:
      type: object
      required: [action]
      properties:
        action:
          type: string
    Notification:
      type: object
      required: [id, userId, type, message, read, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
        message:
          type: string
        read:
          type: boolean
        relatedPostId:
          type: ["integer", "null"]
        relatedUserId:
          type: ["integer", "null"]
        relatedUserName:
          type: ["string", "null"]
        relatedUserAvatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
...