# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Fakebook Social Simulator API
 * OpenAPI spec version: 0.1.0
 */

export interface Comment {
  id: number;
  postId: number;
  userId: number;
  userName: string;
  userAvatarUrl: string;
  content: string;
  /** @nullable */
  parentId?: number | null;
  createdAt: string;
  replies?: Comment[];
}