# lib/api-zod/src/generated/types/friendList.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Fakebook Social Simulator API
 * OpenAPI spec version: 0.1.0
 */
import type { FriendRequest } from './friendRequest';
import type { User } from './user';

export interface FriendList {
  friends: User[];
  pendingRequests: FriendRequest[];
}