# lib/api-zod/src/generated/types/friendRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Fakebook Social Simulator API
 * OpenAPI spec version: 0.1.0
 */

export interface FriendRequest {
  id: number;
  fromUserId: number;
  toUserId: number;
  fromUserName: string;
  fromUserAvatarUrl: string;
  status: string;
  createdAt: string;
}