# lib/api-zod/src/generated/types/postInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Fakebook Social Simulator API
 * OpenAPI spec version: 0.1.0
 */

export interface PostInput {
  /** @minLength 1 */
  content: string;
  imageUrl?: string;
}