# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Fakebook Social Simulator API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  name: string;
  username: string;
  avatarUrl: string;
  coverUrl: string;
  /** @nullable */
  bio: string | null;
  friendCount: number;
  postCount: number;
  isCurrentUser: boolean;
  /** @nullable */
  friendshipStatus?: string | null;
}